backend/requirements.txt
django>=5.0
djangorestframework>=3.14
psycopg2-binary>=2.9
django-cors-headers>=4.3
gunicorn>=21.2
python-dotenv>=1.0
uvicorn>=0.27
